Move Prober to its own subpackage.
This will allow it to be reused by the runtimes.
